+++ XamoomSDK/Classes/Web/XMMWebViewController.m
@@ -16,11 +16,22 @@
 - (void)viewDidLoad {
   [super viewDidLoad];
   
-  [self.navigationController.navigationBar setTintColor:[UIColor whiteColor]];
+  if (_navigationBarColor == nil) {
+    _navigationBarColor = UIColor.whiteColor;
+  }
+  
+  [self.navigationController.navigationBar setTintColor:_navigationBarColor];
   [self.navigationController.navigationBar setBackgroundImage:nil forBarMetrics:UIBarMetricsDefault];
   [self.navigationController.navigationBar setShadowImage:nil];
   [self.navigationController.navigationBar setTranslucent: YES];
   
+  NSString* title = NSLocalizedString(@"webviewcontroller.back", @"");
+  if ([title isEqualToString:@"webviewcontroller.back"]) {
+    title = @"";
+  }
+  
+  self.navigationController.navigationBar.topItem.title = title;
+
   UIWebView *web = [[UIWebView alloc] initWithFrame:CGRectMake(0, 0, self.view.bounds.size.width, self.view.bounds.size.height)];
   NSURL *urlFromString = [NSURL URLWithString:self.url];
   NSURLRequest *request = [NSURLRequest requestWithURL:urlFromString];